requirements.txt
beautifulsoup4==4.12.3
bokeh==3.1.1
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
contourpy==1.1.1
cryptography==42.0.6
greenlet==3.0.3
idna==3.7
Jinja2==3.1.4
joblib==1.4.2
lxml==5.2.2
MarkupSafe==2.1.5
numpy==1.26.4
packaging==24.0
pandas==2.2.3
pillow==10.3.0
pycparser==2.22
PyMySQL==1.1.0
python-dateutil==2.9.0.post0
python-nmap==0.7.1
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
scapy==2.5.0
scikit-learn==1.3.2
scipy==1.14.1
setuptools==68.2.2
six==1.16.0
soupsieve==2.5
SQLAlchemy==2.0.30
threadpoolctl==3.5.0
tornado==6.4
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.2
xyzservices==2024.4.0
